requirements.txt
openai
requests
beautifulsoup4
scikit-learn
flask
regex
pandas
sentence-transformers
faiss-cpu
numpy
spacy==3.7.2
dotenv
keybert
google-auth
google-auth-oauthlib
google-api-python-client
lxml
selenium
#spacy
scrapy
tiktoken
qdrant_client
# venv/bin/activate

# LangChain packages
langchain
langchain_openai
langchain-core
langchain-community
langchain-anthropic
langchain-tavily
langgraph
anthropic
tavily-python